Update Readme.md with more dependencies.
This also moves the `-t babelify` bits to `package.json` so that the
`es2015` presets can be specified. Technically that could go into the
`browserify` bits, but eh.

A future improvement can be to add the dependencies to `package.json` but
I'm super unfamiliar with nodejs so I'll leave that for someone else
:heart: